fixup! feat(random): add @endo/random ChaCha20-based seedable PRNG
Drop the `node:crypto`-backed implementation.  At ChaCha20's 64-byte
block granularity the per-call JS↔native FFI cost of `cipher.update()`
exceeds the inlined pure-JS quarter rounds, so the Node path was
slower, not faster.  The Node fast path was load-bearing for code
size and parity testing rather than performance, and the parity
tests added complexity without protecting any user-visible behavior.

Removed:
- `index-node.js`, `src/random-node.js`, `src/random-pure.js`
- `test/parity.test.js`
- the `"node"` export condition in `package.json`

Folded the former `random-pure.js` wiring directly into `index.js`.
Updated the benchmark to compare just `xorshift128+` vs the pure-JS
`@endo/random`, with a comment recording why the Node path was
considered and dropped.  Updated README and changeset accordingly.
