--- packages/random/index-node.js
@@ -1,1 +1,0 @@
-export { makeRandom } from './src/random-node.js';
